requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.1.0.post1
itsdangerous==2.1.2

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Caching
redis==5.0.1
aioredis==2.0.1

# Monitoring and logging
prometheus-client==0.19.0
structlog==24.1.0
python-json-logger==2.0.7
sentry-sdk==1.39.1

# HTTP client
httpx==0.26.0
requests==2.31.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-env==1.1.3

# Development tools
black==24.1.0
isort==5.13.2
mypy==1.8.0
flake8==7.0.0
pre-commit==3.6.0

# Utilities
tenacity==8.2.3
PyYAML==6.0.1
